% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{bic}
\alias{bic}
\title{BIC scorer function}
\usage{
bic(y, y_pred, p)
}
\arguments{
\item{y}{True target variable(s) - array-like of shape = (n_samples) or (n_samples, n_outputs)}

\item{y_pred}{is a list of numbers and p which is an int}

\item{p}{Number of predictive variable(s) used in the model - int}
}
\value{
double
}
\description{
BIC stands for Bayesian Information Criterion.
Like AIC, it also estimates the quality of a model.
When fitting models, it is possible to increase model fitness by adding more parameters.
Doing this may results in model overfit.
Both AIC and BIC helps to resolve this problem by using a penalty term for the number of parameters in the model.
This term is bigger in BIC than in AIC.
}
\details{
BIC = n X log(residual sum of squares/n) + K X log(n)
}
\examples{
library(regscoreR)
y <-c(1,2,3,4)
y_pred <- c(5,6,7,8)
p <- 3
regscoreR::bic(y, y_pred, p)

}
\author{
Simran Sethi
}
